requirements.txt
supabase==2.4.5
google-genai>=1.0.0
pyjwt==2.10.1
